# terminus_openapi.yaml
openapi: 3.0.0
info:
  title: Terminus API
  version: 1.0.0
  description: API specification for the Terminus e-ink display server.
servers:
  - url: https://localhost:2443
    description: Local development server
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceMac:
      type: apiKey
      in: header
      name: ID
  schemas:
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        wifi_status:
          type: string
        created_at:
          type: integer
        sleep_duration:
          type: integer
        refresh_rate:
          type: integer
        free_heap_size:
          type: integer
        max_alloc_size:
          type: integer
        source_path:
          type: string
        wake_reason:
          type: string
        firmware_version:
          type: string
        retry:
          type: integer
        battery_voltage:
          type: number
        source_line:
          type: integer
        special_function:
          type: string
        wifi_signal:
          type: integer
paths:
...
  /api/display:
    get:
      summary: Get current display content for a device
      tags:
        - Firmware
      security:
        - deviceMac: []
      parameters:
        - in: header
          name: ID
          required: true
          schema:
            type: string
          description: Device MAC address
        - in: header
          name: BATTERY_VOLTAGE
          schema:
            type: number
        - in: header
          name: FW_VERSION
          schema:
            type: string
        - in: header
          name: RSSI
          schema:
            type: integer
        # Other headers can be added as needed based on the list
      responses:
        '200':
          description: Successful response with image URL and settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  firmware_url:
                    type: string
                    nullable: true
                  firmware_version:
                    type: string
                    nullable: true
                  image_url:
                    type: string
                  image_url_timeout:
                    type: integer
                  maximum_compatibility:
                    type: boolean
                  refresh_rate:
                    type: integer
                  reset_firmware:
                    type: boolean
                  special_function:
                    type: string
                  temperature_profile:
                    type: string
                  touchbar_mode:
                    type: string
                  update_firmware:
                    type: boolean
  /api/log:
    post:
      summary: Push device logs to the server
      tags:
        - Firmware
      security:
        - deviceMac: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logs:
                  type: array
                  items:
                    $ref: '#/components/schemas/LogEntry'
      responses:
        '204':
          description: Logs successfully recorded
        '401':
          description: Unauthorized
  /api/setup:
    get:
      summary: Provision a new device
      tags:
        - Firmware
      security:
        - deviceMac: []
      responses:
        '200':
          description: Device provisioned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  image_url:
                    type: string
                  message:
                    type: string
...